feat: Add react-icons package and update dependencies
This commit adds the react-icons package to the project and updates the dependencies in the package.json and package-lock.json files. It also includes the addition of the @types/react-icons package as a dev dependency. The react-icons package is now at version 5.3.0 and the @types/react-icons package is at version 2.2.7.

The purpose of this change is to enable the usage of icons in the application using the react-icons library.

Closes #<issue_number>